Remove std/multipart (#3647)
since it overlaps with std/mime/multipart
