--- std/multipart/formfile_test.ts
@@ -1,32 +1,0 @@
-// Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
-import { test } from "../testing/mod.ts";
-import { assertEquals } from "../testing/asserts.ts";
-import { isFormFile } from "./formfile.ts";
-
-test(function multipartIsFormFile(): void {
-  assertEquals(
-    isFormFile({
-      filename: "foo",
-      type: "application/json"
-    }),
-    true
-  );
-  assertEquals(
-    isFormFile({
-      filename: "foo"
-    }),
-    false
-  );
-});
-
-test(function isFormFileShouldNotThrow(): void {
-  assertEquals(
-    isFormFile({
-      filename: "foo",
-      type: "application/json",
-      hasOwnProperty: "bar"
-    }),
-    true
-  );
-  assertEquals(isFormFile(Object.create(null)), false);
-});
