--- std/multipart/mod.ts
@@ -1,1 +1,0 @@
-export * from "./formfile.ts";
